+++ src/core/project/index.ts
@@ -1,5 +1,6 @@
 export { CalmProject } from './CalmProject';
 export { createProject } from './create';
+export { deleteProject } from './delete';
 export { getProgram, getProject } from './get';
 export { listPrograms, listProjects } from './list';
 export { listProjectTeamMembers, listProjectTimeboxes } from './nested';
@@ -9,4 +10,6 @@ export type {
   IProject,
   ITeamMember,
   ITimebox,
+  IUpdateProjectParams,
 } from './types';
+export { updateProject } from './update';