+++ src/core/testCase/index.ts
@@ -1,17 +1,50 @@
-export { createTestAction, listTestActions } from './actions';
-export { createTestActivity, listTestActivities } from './activities';
+export {
+  createActivityAction,
+  createTestAction,
+  deleteTestAction,
+  getTestAction,
+  listTestActions,
+  updateTestAction,
+} from './actions';
+export {
+  createTestActivity,
+  createTestCaseActivity,
+  deleteTestActivity,
+  getTestActivity,
+  listTestActivities,
+  listTestCaseActivities,
+  updateTestActivity,
+} from './activities';
+export {
+  createActivityApplication,
+  listActivityApplications,
+} from './applications';
 export { CalmTestCase } from './CalmTestCase';
 export { createTestCase } from './create';
-export { deleteTestCase } from './delete';
+export { deleteTestCase, forceDeleteTestCase } from './delete';
 export { getTestCase } from './get';
 export { listTestCases } from './list';
+export { createTestCaseReference, listTestCaseReferences } from './references';
+export {
+  listTestCaseTagAssignments,
+  listTestTagAssignments,
+} from './tagAssignments';
 export type {
+  ICreateActivityActionParams,
   ICreateTestActionParams,
   ICreateTestActivityParams,
+  ICreateTestApplicationParams,
+  ICreateTestCaseActivityParams,
   ICreateTestCaseParams,
+  ICreateTestReferenceParams,
   ITestAction,
   ITestActivity,
+  ITestApplication,
   ITestCase,
+  ITestReference,
+  ITestTagAssignment,
+  IUpdateTestActionParams,
+  IUpdateTestActivityParams,
   IUpdateTestCaseParams,
 } from './types';
 export { updateTestCase } from './update';